+++ KeyboardKit/Views/KeyboardCollectionView.swift
@@ -1,0 +1,28 @@
+//
+//  KeyboardCollectionView.swift
+//  KeyboardKit
+//
+//  Created by Daniel Saidi on 2019-04-23.
+//  Copyright © 2019 Daniel Saidi. All rights reserved.
+//
+
+import UIKit
+
+open class KeyboardCollectionView: UICollectionView {
+
+    public override init(frame: CGRect, collectionViewLayout layout: UICollectionViewLayout) {
+        super.init(frame: frame, collectionViewLayout: layout)
+        setup()
+    }
+    
+    public required init?(coder aDecoder: NSCoder) {
+        super.init(coder: aDecoder)
+        setup()
+    }
+    
+    open func setup() {
+        bounces = false
+        isPagingEnabled = true
+        backgroundColor = .clear
+    }
+}